fixup: some cleanups
